feat: pickup feature, add and update domain models, integrate between pages, adjust waste details page

+++ Roambank/Models/Waste.swift
@@ -10,9 +10,9 @@ import Foundation
 struct Waste: Identifiable {
     let id: UUID
     let wasteType: WasteType
-    var berat: Int
+    var berat: Double
     
-    init(id: UUID = UUID(), wasteType: WasteType, berat: Int) {
+    init(id: UUID = UUID(), wasteType: WasteType, berat: Double) {
         self.id = id
         self.wasteType = wasteType
         self.berat = berat
